pyproject.toml
[tool.poetry]
name = "cashly-ai"
version = "2.0.0"
description = "Cashly AI Financial Service - FastAPI"
authors = ["Cashly Team"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
fastapi = "^0.104.1"
uvicorn = { extras = ["standard"], version = "^0.34.2" }
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.41"
asyncpg = "^0.30.0"
pgvector = "^0.4.1"
httpx = {extras = ["http2"], version = "^0.28.1"}
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-multipart = "^0.0.6"
python-dotenv = "^1.0.1"
openai = "^1.82.0"
pandas = "^2.2.3"
numpy = "^2.2.4"
scikit-learn = "^1.6.1"
joblib = "^1.4.2"
tiktoken = "^0.9.0"
aiohttp = "^3.12.3"
psutil = "^7.0.0"
prophet = "^1.1.7"
matplotlib = "^3.10.3"
statsmodels = "^0.14.4"
spacy = "^3.8.7"
transformers = "^4.52.4"
huggingface-hub = "^0.32.3"
torch = "^2.7.0"
requests = "^2.32.3"
psycopg2-binary = "^2.9.10"
aiofiles = "^24.1.0"
boto3 = "^1.38.28"
mlflow = "^2.22.0"
seaborn = "^0.13.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.5.0"
pytest-mock = "^3.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=app --cov-report=term-missing"